+++ lib/settings/view/widgets/widgets.dart
@@ -1,1 +1,2 @@
 export 'theme_section.dart';
+export 'about_section.dart';