+++ lib/shared/background_worker.dart
@@ -1,11 +1,9 @@
-import 'package:awesome_snackbar_content/awesome_snackbar_content.dart';
-import 'package:qruto_budget/backup/cubit/backup_cubit.dart';
-import 'package:qruto_budget/constants/changelog.dart';
-import 'package:qruto_budget/utils/theme/cubit/theme_cubit.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:in_app_update/in_app_update.dart';
 import 'package:package_info_plus/package_info_plus.dart';
+import 'package:qruto_budget/backup/cubit/backup_cubit.dart';
+import 'package:qruto_budget/shared/shared_functions.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
 class BackgroundWorker {
@@ -31,7 +29,7 @@ class BackgroundWorker {
     if (currentVersion != lastVersion) {
       // Show the bottom sheet
       WidgetsBinding.instance.addPostFrameCallback((_) {
-        _showWhatsNewSheet(context);
+        SharedFunctions.showWhatsNewSheet(context);
       });
 
       // Update stored version
@@ -49,21 +47,9 @@ class BackgroundWorker {
     if (dateLastAutoBackup == null ||
         DateTime.now().difference(dateLastAutoBackup).inHours > 24) {
       final success = await context.read<BackupCubit>().autoBackup();
-      if(success){
-        final snackBar = SnackBar(
-          elevation: 0,
-          behavior: SnackBarBehavior.floating,
-          backgroundColor: Colors.transparent,
-          content: AwesomeSnackbarContent(
-            title: 'Auto Backup !',
-            message: 'Auto Backup has been uploaded.',
-            contentType: ContentType.success,
-          ),
-        );
-
-        ScaffoldMessenger.of(context)
-          ..hideCurrentSnackBar()
-          ..showSnackBar(snackBar);
+      if (success) {
+        SharedFunctions.showSnackbar(
+            context, true, 'Auto Backup !', 'Auto Backup has been uploaded.');
       }
     }
   }
@@ -116,150 +102,6 @@ class BackgroundWorker {
           ],
         ),
       ),
-    );
-  }
-
-  static void _showWhatsNewSheet(BuildContext context) {
-    showModalBottomSheet<void>(
-      isScrollControlled: true,
-      context: context,
-      builder: (BuildContext context) {
-        final prColor = context.read<ThemeCubit>().state.primaryColor;
-        return Container(
-          height: 600,
-          width: double.infinity,
-          padding: EdgeInsets.symmetric(horizontal: 0, vertical: 40),
-          child: SingleChildScrollView(
-            child: Padding(
-              padding:
-                  const EdgeInsets.symmetric(horizontal: 40.0, vertical: 0),
-              child: Column(
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: [
-                  Text("What's New",
-                      style:
-                          TextStyle(fontWeight: FontWeight.bold, fontSize: 30)),
-                  SizedBox(height: 20),
-                  Divider(),
-                  SizedBox(
-                    height: 15,
-                  ),
-                  Column(
-                    crossAxisAlignment: CrossAxisAlignment.start,
-                    children: [
-                      ...changelog.map((item) {
-                        final isTitle =
-                            (item['titles'] as List<dynamic>).isNotEmpty;
-                        final isAdded =
-                            (item['added'] as List<dynamic>).isNotEmpty;
-                        final isFixed =
-                            (item['fixed'] as List<dynamic>).isNotEmpty;
-                        return Column(
-                          crossAxisAlignment: CrossAxisAlignment.start,
-                          children: [
-                            Container(
-                              padding: EdgeInsets.symmetric(
-                                  vertical: 0, horizontal: 8),
-                              // Padding inside the box
-                              decoration: BoxDecoration(
-                                color: prColor[100], // Background color
-                                borderRadius:
-                                    BorderRadius.circular(8), // Rounded corners
-                              ),
-                              child: Text(
-                                item['version'] as String,
-                                // Assuming `item` has a `version` field
-                                style: TextStyle(
-                                    fontWeight: FontWeight.bold, fontSize: 22),
-                              ),
-                            ),
-                            Text(
-                              item['date'] as String,
-                              // Assuming `item` has a `version` field
-                              style: TextStyle(fontSize: 12),
-                            ),
-                            SizedBox(
-                              height: 15,
-                            ),
-                            isTitle
-                                ? RichText(
-                                    text: TextSpan(
-                                        style: TextStyle(
-                                            height: 1.5,
-                                            fontSize: 20,
-                                            color: Colors.black),
-                                        children:
-                                            (item['titles'] as List<String>)
-                                                .map<TextSpan>((change) =>
-                                                    TextSpan(text: '$change\n'))
-                                                .toList()),
-                                  )
-                                : Container(),
-                            isAdded
-                                ? Column(
-                                    crossAxisAlignment:
-                                        CrossAxisAlignment.start,
-                                    children: [
-                                      Text('Added:',
-                                          // Assuming `item` has a `version` field
-                                          style: TextStyle(
-                                              fontSize: 20,
-                                              fontWeight: FontWeight.bold)),
-                                      RichText(
-                                          text: TextSpan(
-                                              style: TextStyle(
-                                                  fontSize: 20,
-                                                  color: Colors.black,
-                                                  height: 1.5),
-                                              children: (item['added']
-                                                      as List<String>)
-                                                  .map<TextSpan>((change) =>
-                                                      TextSpan(
-                                                          text: ' - $change\n'))
-                                                  .toList())),
-                                    ],
-                                  )
-                                : Container(),
-                            isFixed
-                                ? Column(
-                                    crossAxisAlignment:
-                                        CrossAxisAlignment.start,
-                                    children: [
-                                      Text(
-                                        'Fixed:',
-                                        // Assuming `item` has a `version` field
-                                        style: TextStyle(
-                                            fontSize: 20,
-                                            fontWeight: FontWeight.bold),
-                                      ),
-                                      RichText(
-                                          text: TextSpan(
-                                              style: TextStyle(
-                                                  height: 1.5,
-                                                  fontSize: 20,
-                                                  color: Colors.black),
-                                              children: (item['fixed']
-                                                      as List<String>)
-                                                  .map<TextSpan>((change) =>
-                                                      TextSpan(
-                                                          text: ' - $change\n'))
-                                                  .toList()))
-                                    ],
-                                  )
-                                : Container(),
-                            Divider(),
-                            SizedBox(height: 15),
-                          ],
-                        );
-                      }),
-                    ],
-                  )
-                ],
-              ),
-            ),
-          ),
-        );
-      },
     );
   }
 }