+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [4.5.0] - 03-29-2025
+### Added
+- Select a custom range in the Summary
+- Predefined income or expense categories
+- About section in the Settings page
+### Fixed
+- Category or Account can't be deleted if they contain any transaction
+
 ## [4.4.0] - 03-26-2025
 ### Added
 - Auto backup to Google Drive